add voicemail

+++ lib/t/matches.ex
@@ -10,6 +10,7 @@ defmodule T.Matches do
   alias T.Matches.{Match, Like, Timeslot, MatchEvent, ExpiredMatch, MatchContact, ArchivedMatch}
   alias T.Feeds.{FeedProfile, LiveSession}
   alias T.Accounts.{Profile, UserSettings}
+  alias T.Calls.Voicemail
   alias T.PushNotifications.DispatchJob
   alias T.Bot
 
@@ -20,7 +21,14 @@ defmodule T.Matches do
   @pubsub T.PubSub
   @topic "__m"
 
-  @match_ttl 604_800
+  @seven_days 7 * 24 * 60 * 60
+  @two_days 2 * 24 * 60 * 60
+
+  @doc "Time-to-live for a match without life-prolonging events like calls, meetings, voice-messages"
+  def match_ttl, do: @seven_days
+
+  @doc "Time-to-live for a match before voicemail has been exchanged both ways"
+  def pre_voicemail_ttl, do: @two_days
 
   defp pubsub_user_topic(user_id) when is_binary(user_id) do
     @topic <> ":u:" <> String.downcase(user_id)
@@ -73,32 +81,24 @@ defmodule T.Matches do
   end
 
   defp maybe_notify_match(
-         %Match{id: match_id} = match,
+         %Match{id: match_id, inserted_at: inserted_at} = match,
          [by_user_id_settings, user_id_settings],
          by_user_id,
          user_id
        ) do
-    broadcast_from_for_user(
-      by_user_id,
-      {__MODULE__, :matched,
-       %{
-         id: match_id,
-         expiration_date: expiration_date(match),
-         mate: user_id,
-         audio_only: user_id_settings
-       }}
-    )
+    expiration_date = pre_voicemail_expiration_date(match)
 
-    broadcast_for_user(
-      user_id,
-      {__MODULE__, :matched,
-       %{
-         id: match_id,
-         expiration_date: expiration_date(match),
-         mate: by_user_id,
-         audio_only: by_user_id_settings
-       }}
-    )
+    common = %{
+      id: match_id,
+      inserted_at: inserted_at,
+      expiration_date: expiration_date
+    }
+
+    by_user_payload = Map.merge(common, %{mate: user_id, audio_only: user_id_settings})
+    user_payload = Map.merge(common, %{mate: by_user_id, audio_only: by_user_id_settings})
+
+    broadcast_from_for_user(by_user_id, {__MODULE__, :matched, by_user_payload})
+    broadcast_for_user(user_id, {__MODULE__, :matched, user_payload})
   end
 
   defp maybe_notify_match(nil, _settings, _by_user_id, _user_id), do: :ok
@@ -318,31 +318,50 @@ defmodule T.Matches do
     |> where([match: m], m.user_id_1 == ^user_id or m.user_id_2 == ^user_id)
     |> where([match: m], m.id not in subquery(archived_match_ids_q(user_id)))
     |> order_by(desc: :inserted_at)
-    |> join(:left, [match: m], t in Timeslot, as: :t, on: m.id == t.match_id)
-    |> join(:left, [match: m], c in MatchContact, as: :c, on: m.id == c.match_id)
-    |> select([match: m, undying_event: e, t: t, c: c], {m, t, c, e.timestamp})
+    |> join(:left, [m], t in assoc(m, :timeslot), as: :t)
+    |> join(:left, [m], c in assoc(m, :contact), as: :c)
+    |> join(:left, [m], v in assoc(m, :voicemail), as: :v)
+    # TODO aggregate voicemail
+    |> preload([t: t, c: c, v: v], timeslot: t, contact: c, voicemail: v)
+    |> select([match: m, undying_event: e], {m, e.timestamp})
     |> Repo.all()
-    |> Enum.map(fn {match, timeslot, contact, undying_event_timestamp} ->
+    |> Enum.map(fn {match, undying_event_timestamp} ->
       expiration_date =
-        unless undying_event_timestamp do
-          expiration_date(match)
+        cond do
+          undying_event_timestamp -> nil
+          match.exchanged_voicemail -> post_voicemail_expiration_date(match)
+          true -> pre_voicemail_expiration_date(match)
         end
 
       %Match{
         match
-        | timeslot: timeslot,
-          contact: contact,
-          expiration_date: expiration_date,
-          interaction: timeslot || contact
+        | expiration_date: expiration_date,
+          interaction: interaction(match)
       }
     end)
     |> preload_match_profiles(user_id)
   end
 
-  defp expiration_date(%Match{inserted_at: inserted_at}) do
+  defp interaction(%Match{timeslot: %Timeslot{} = timeslot}), do: timeslot
+  defp interaction(%Match{contact: %MatchContact{} = contact}), do: contact
+
+  # TODO order voicemail in db
+  defp interaction(%Match{voicemail: [%Voicemail{} | _rest] = voicemail}) do
+    Enum.sort_by(voicemail, & &1.id)
+  end
+
+  defp interaction(%Match{}), do: nil
+
+  defp post_voicemail_expiration_date(%Match{inserted_at: inserted_at}) do
     inserted_at
     |> DateTime.from_naive!("Etc/UTC")
-    |> DateTime.add(@match_ttl)
+    |> DateTime.add(match_ttl())
+  end
+
+  defp pre_voicemail_expiration_date(%Match{inserted_at: inserted_at}) do
+    inserted_at
+    |> DateTime.from_naive!("Etc/UTC")
+    |> DateTime.add(pre_voicemail_ttl())
   end
 
   defp archived_match_ids_q(user_id) do
@@ -396,6 +415,7 @@ defmodule T.Matches do
     Logger.warn("#{by_user_id} unmatches match-id=#{match_id}")
 
     Multi.new()
+    |> delete_voicemail(match_id)
     |> Multi.run(:unmatch, fn _repo, _changes ->
       Match
       |> where(id: ^match_id)
@@ -426,14 +446,24 @@ defmodule T.Matches do
     [user_id_1, user_id_2] = Enum.sort([by_user_id, mate])
 
     Multi.run(multi, :unmatch, fn _repo, _changes ->
-      Match
-      |> where(user_id_1: ^user_id_1)
-      |> where(user_id_2: ^user_id_2)
-      |> select([m], m.id)
-      |> Repo.delete_all()
-      |> case do
-        {1, [match_id]} -> {:ok, fn -> notify_unmatch(by_user_id, mate, match_id) end}
-        {0, _} -> {:ok, nil}
+      match_id =
+        Match
+        |> where(user_id_1: ^user_id_1)
+        |> where(user_id_2: ^user_id_2)
+        |> select([m], m.id)
+        |> Repo.one()
+
+      if match_id do
+        :ok = T.Calls.voicemail_delete_all(match_id)
+
+        {1, _} =
+          Match
+          |> where(id: ^match_id)
+          |> Repo.delete_all()
+
+        {:ok, fn -> notify_unmatch(by_user_id, mate, match_id) end}
+      else
+        {:ok, nil}
       end
     end)
   end
@@ -453,14 +483,24 @@ defmodule T.Matches do
 
     Multi.new()
     |> Multi.run(:unmatch, fn _repo, _changes ->
-      Match
-      |> where(user_id_1: ^uid1)
-      |> where(user_id_2: ^uid2)
-      |> select([m], %{id: m.id, users: [m.user_id_1, m.user_id_2]})
-      |> Repo.delete_all()
-      |> case do
-        {1, [match]} -> {:ok, match}
-        {0, _} -> {:error, :match_not_found}
+      match_id =
+        Match
+        |> where(user_id_1: ^uid1)
+        |> where(user_id_2: ^uid2)
+        |> select([m], m.id)
+        |> Repo.one()
+
+      if match_id do
+        :ok = T.Calls.voicemail_delete_all(match_id)
+
+        {1, _} =
+          Match
+          |> where(id: ^match_id)
+          |> Repo.delete_all()
+
+        {:ok, %{id: match_id, users: [uid1, uid2]}}
+      else
+        {:error, :match_not_found}
       end
     end)
     |> delete_likes()
@@ -496,6 +536,7 @@ defmodule T.Matches do
     Bot.async_post_message(m)
 
     Multi.new()
+    |> delete_voicemail(match_id)
     |> Multi.run(:unmatch, fn _repo, _changes ->
       Match
       |> where(user_id_1: ^user_id_1)
@@ -592,6 +633,16 @@ defmodule T.Matches do
     end)
   end
 
+  # note this needs to run before match delete
+  # otherwise DB records would be CASCADE deleted already
+  # and we wouldn't know s3_keys to delete
+  defp delete_voicemail(multi, match_id) do
+    Multi.run(multi, :voicemail, fn _repo, _changes ->
+      :ok = T.Calls.voicemail_delete_all(match_id)
+      {:ok, nil}
+    end)
+  end
+
   defp insert_expired_match(multi) do
     Multi.insert_all(multi, :insert_expired_match, ExpiredMatch, fn %{unmatch: unmatch} ->
       %{id: match_id, users: [user_id_1, user_id_2]} = unmatch
@@ -669,6 +720,7 @@ defmodule T.Matches do
 
     Multi.new()
     |> Multi.delete_all(:contacts, where(MatchContact, match_id: ^match_id))
+    |> delete_voicemail(match_id)
     |> Multi.insert(:timeslot, changeset, conflict_opts)
     |> Multi.insert(:match_event, %MatchEvent{
       timestamp: DateTime.truncate(DateTime.utc_now(), :second),
@@ -882,6 +934,7 @@ defmodule T.Matches do
 
     Multi.new()
     |> Multi.delete_all(:timeslots, where(Timeslot, match_id: ^match_id))
+    |> delete_voicemail(match_id)
     |> Multi.insert(:match_contact, changeset, conflict_opts)
     |> Multi.insert(:match_event, %MatchEvent{
       timestamp: now,
@@ -1083,7 +1136,7 @@ defmodule T.Matches do
 
   # Expired Matches
 
-  def insert_match_event(match_id, event) do
+  defp insert_match_event(match_id, event) do
     Repo.insert(%MatchEvent{
       timestamp: DateTime.truncate(DateTime.utc_now(), :second),
       match_id: match_id,
@@ -1091,36 +1144,69 @@ defmodule T.Matches do
     })
   end
 
-  def match_expired_check(reference \\ DateTime.utc_now()) do
-    expiration_date = DateTime.add(reference, -@match_ttl)
-
-    expiring_matches_q()
-    |> where([m, c], m.inserted_at < ^expiration_date)
-    |> select([m], {m.id, m.user_id_1, m.user_id_2})
-    |> T.Repo.all()
-    |> Enum.map(fn {match_id, user_id_1, user_id_2} ->
+  def expiration_prune(reference \\ DateTime.utc_now()) do
+    reference
+    |> expiration_list_expired_matches()
+    |> Enum.each(fn %{id: match_id, user_id_1: user_id_1, user_id_2: user_id_2} ->
       expire_match(match_id, user_id_1, user_id_2)
     end)
   end
 
-  def match_soon_to_expire_check(reference \\ DateTime.utc_now()) do
-    day_before_expiration = reference |> DateTime.add(-@match_ttl) |> DateTime.add(24 * 3600)
-    almost_day_before_expiration = DateTime.add(day_before_expiration, -60)
+  def expiration_list_expired_matches(reference \\ DateTime.utc_now()) do
+    post_voicemail_expiration_date = DateTime.add(reference, -match_ttl())
+    pre_voicemail_expiration_date = DateTime.add(reference, -pre_voicemail_ttl())
 
     expiring_matches_q()
-    |> where([m], m.inserted_at <= ^day_before_expiration)
-    # TODO can result in duplicates with more than one worker
-    |> where([m], m.inserted_at > ^almost_day_before_expiration)
-    |> select([m], m.id)
-    |> T.Repo.all()
-    |> Enum.map(fn match_id ->
-      schedule_match_about_to_expire(match_id)
-    end)
+    |> where(
+      [match: m],
+      (not m.exchanged_voicemail and m.inserted_at < ^pre_voicemail_expiration_date) or
+        m.inserted_at < ^post_voicemail_expiration_date
+    )
+    |> select([match: m], map(m, [:id, :user_id_1, :user_id_2]))
+    |> Repo.all()
   end
 
-  defp schedule_match_about_to_expire(match_id) do
-    job = DispatchJob.new(%{"type" => "match_about_to_expire", "match_id" => match_id})
-    Oban.insert(job)
+  def expiration_notify_soon_to_expire(reference \\ DateTime.utc_now()) do
+    expiration_list_soon_to_expire(reference)
+    |> Enum.map(fn match_id ->
+      DispatchJob.new(%{"type" => "match_about_to_expire", "match_id" => match_id})
+    end)
+    |> Oban.insert_all()
+  end
+
+  defp expiration_pre_voicemail_notification_interval(reference) do
+    to =
+      reference
+      |> DateTime.add(-pre_voicemail_ttl())
+      |> DateTime.add(_3_hours = 3 * 3600)
+
+    from = DateTime.add(to, -60)
+    {from, to}
+  end
+
+  defp expiration_post_voicemail_notification_interval(reference) do
+    to =
+      reference
+      |> DateTime.add(-match_ttl())
+      |> DateTime.add(_24_hours = 24 * 3600)
+
+    from = DateTime.add(to, -60)
+    {from, to}
+  end
+
+  def expiration_list_soon_to_expire(reference \\ DateTime.utc_now()) do
+    {from1, to1} = expiration_pre_voicemail_notification_interval(reference)
+    {from2, to2} = expiration_post_voicemail_notification_interval(reference)
+
+    expiring_matches_q()
+    # TODO can result in duplicates with more than one worker
+    |> where(
+      [match: m],
+      (not m.exchanged_voicemail and (m.inserted_at > ^from1 and m.inserted_at <= ^to1)) or
+        (m.inserted_at > ^from2 and m.inserted_at <= ^to2)
+    )
+    |> select([m], m.id)
+    |> Repo.all()
   end
 
   def delete_expired_match(match_id, by_user_id) do
@@ -1140,19 +1226,14 @@ defmodule T.Matches do
   end
 
   defp matches_with_undying_events_q(query \\ named_match_q()) do
-    join(
-      query,
-      :left_lateral,
-      [m, e],
-      c in subquery(
-        MatchEvent
-        |> where(match_id: parent_as(:match).id)
-        |> where([e], e.event == "call_start" or e.event == "meeting_report")
-        |> select([e], e.timestamp)
-        |> limit(1)
-      ),
-      as: :undying_event
-    )
+    undying_events_q =
+      MatchEvent
+      |> where(match_id: parent_as(:match).id)
+      |> where([e], e.event == "call_start" or e.event == "meeting_report")
+      |> select([e], e.timestamp)
+      |> limit(1)
+
+    join(query, :left_lateral, [m], e in subquery(undying_events_q), as: :undying_event)
   end
 
   # Contact Exchange
